moved service deployment logic from service pkg to api pkg
Service type will be simplifed by moving its methods to the container Manager that will be introduced in future. Manager will be about managing the containers related to service. thus, it cannot have logic related to building a Docker image (service deployment) and service pkg cannot keep this logic as well. as a result, this logic is moved to the api pkg.