introduce the first version of Workflow Object & VM
Secret can be carried as an execution tag. This Workflow VM does not depend on internal packages so, it's very lightweight, simple and easy to test.

VM uses Process(Event) method to receive(I of I/O) new event(/result) and uses ExecutionRequests channel to send(O of I/O) execution requests for new or in progress workflow lifecycles that match with state of Event.